#39 Created registration endpoint so that the form can send data to the database
I added the register route and controller so that when the register form is submitted on the registration page, the data will be sent to the database so that it can be entered. Since the registration page isn't actually created yet, that names I used to parse may need to be updated.